% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{cnrm}
\alias{bcc}
\alias{bnu}
\alias{cccma}
\alias{cmcc}
\alias{cnrm}
\alias{csiro}
\alias{fio}
\alias{gfdl}
\alias{giss}
\alias{iap}
\alias{ichec}
\alias{ingv}
\alias{inm}
\alias{ipsl}
\alias{miroc}
\alias{mohc}
\alias{mpim}
\alias{mri}
\alias{ncar}
\alias{ncc}
\alias{obs}
\title{European and global summer mean temperatures from 1850 to 2100}
\format{A data frame}
\usage{
cnrm

bcc

bnu

cccma

cmcc

cnrm

csiro

fio

gfdl

giss

iap

ichec

ingv

inm

ipsl

miroc

mohc

mpim

mri

ncar

ncc

obs
}
\description{
European and global summer temperature from runs of the CMIP5 simulations
climate simulations. Summer corresponds to the months JJA. First part of the simulations are historical runs and afterwards the scenario RCP8.5 is used.
}
\details{
\itemize{
  \item year. year of the anomalies
  \item eur_tas. European summer temperature anomalies (K)
  \item gbl_tas. Global summer temperature anomalies (K)
  \item run. the run number
  \item type. the type of run: hist, rcp, nat(historical nat)
  \item hant. wether the run is made of historical nat simulations
}
}
\keyword{data}
\keyword{datasets}

