% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_far_simple.r
\name{compute_far_simple}
\alias{compute_far_simple}
\title{function to compute the FAR}
\usage{
compute_far_simple(mdata, y = "y", x = "x", time = "time", xp = 1.6,
  R = 3, stat_model = gauss_fit, ci_p = 0.9, ...)
}
\arguments{
\item{mdata, }{a data.frame containing the variables whose names are given by x, y and time}

\item{y, }{the name of variable that will be used as the variable of interest y}

\item{x, }{the name of variable that will be used as the covariate x}

\item{time, }{the name of variable that will be used as the as the time variable}

\item{xp, }{the threshold used to define the FAR}

\item{R, }{the number of bootstrap samples}

\item{stat_model}{the statistical model to explain y in function of x, either
gauss_fit, gev_fit, or gpd_fit from the FARg package}

\item{ci_p}{the level of the confidence intervals}

\item{...}{additional arguments if required by the stat_model function}
}
\description{
\code{compute_far_simple} Function to compute the FAR from annual time-series
}
\examples{
library(FARg)
model <- "cnrm"
#load data from the package
data(list=model)
# formating data, e.g passing from temperature to anomalie, keep only hist
# and rcp runs
mdata <- format_data(get(model))
if(model != "obs") mdata <- select_continuous_run(mdata)
ans <- compute_far_simple.default(mdata,
 y="eur_tas", x="gbl_tas", time="year",
  xp=1.6, stat_model=gauss_fit, ci_p=0.9)
}
