% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{continue_with_constrained_far}
\alias{continue_with_constrained_far}
\title{Compute the FAR from observations using constraints}
\usage{
continue_with_constrained_far(merged_res)
}
\arguments{
\item{merged_res}{a result from the compute_forall function. the dataset
"obs" should have been treated in the compute_forall function.}
}
\value{
return a dataframe with the confidence intervals for the constrained FAR and all
the other computed parameters (e.g, p and q) at each time.
}
\description{
\code{compute_forall} Compute the FAR from the observations using the GCM to
constrain the evolution of the statistical distribution of the variable of
interest y. The relationship between the variable y and the covariate x in the
observational should be in the observation the same as in the GCM.
}
\examples{
# compute the FAR for the CNRM and the IPSL GCMs using a  gam decomposition
# and a gaussian fit with only three bootstrap samples
ans <- compute_forall(compute_far.default, stat_model=gauss_fit, p=0.01, R=3)
ans_cstr <- continue_with_constrained_far(ans) 
summary_plot(ans_cstr)
}
