% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fun.r
\name{default_plot}
\alias{default_plot}
\title{\code{default_plot} Confidence interval plot for results from the compute_far functions}
\usage{
default_plot(ic_far)
}
\arguments{
\item{ic_far}{a data_frame contaning estimates and their confidence
intervales produced by the compute_far functions, i.e. the ic_far element of the returned list of the compute_far functions}
}
\description{
\code{default_plot} Confidence interval plot for results from the compute_far functions
}
\examples{
library(FARg)
model <- "cnrm"
#load data from the package
data(list=model)
# formating data, e.g passing from temperature to anomalie, keep only hist
# and rcp runs
mdata <- format_data(get(model))
if(model != "obs") mdata <- select_continuous_run(mdata)
ans <- compute_far_simple.default(mdata,
 y="eur_tas", x="gbl_tas", time="year",
  xp=1.6, stat_model=gauss_fit, ci_p=0.9)
ic_x <- subset(ans$ic_far, param \%in\% c("x_all", "x_ant", "x_nat"))
default_plot(ic_x)
}
