% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_x.r
\name{dx.gam_allnat}
\alias{dx.gam_allnat}
\alias{dx.raw}
\alias{dx.lm_gno}
\alias{dx.gam_gno}
\alias{dx.ebm_fit}
\title{Function to decompose the covariate x into an ALL, an ANT and a NAT
component}
\usage{
dx.gam_allnat(bsamples, bindexes, y = "y", x = "x", ant = "ant",
  nat = "nat", time = "time", hnat = "hnat", knots = NULL,
  fixed = FALSE, correct_ant_bias = TRUE, pre_ind = c(1850, 1879),
  reuse_ant_bias = FALSE)

dx.raw(bsamples, bindexes, y = "y", all = "all", ant = "ant",
  nat = "nat", time = "time", hnat = "hnat")

dx.lm_gno(bsamples, bindexes, y = "y", x = "x", ghg = "ghg",
  nat = "nat", others = "others", time = "time", hnat = "hnat",
  correct_ant_bias = TRUE, pre_ind = c(1850, 1879),
  reuse_ant_bias = FALSE)

dx.gam_gno(bsamples, bindexes, y = "y", x = "x", ghg = "ghg",
  nat = "nat", others = "others", time = "time", hnat = "hnat",
  correct_ant_bias = TRUE, pre_ind = c(1850, 1879),
  reuse_ant_bias = FALSE)

dx.ebm_fit(bsamples, bindexes, y = "y", x = "x", time = "time",
  hnat = "hnat")
}
\arguments{
\item{bsamples}{: a list of data.frame with the original dataset merged with the
bootstrap EBM simulations}

\item{bindexes}{: a list of vectors of indexes indicating the rows that will be
used for the bootstrap}

\item{y}{the name of the variable of interest}

\item{x}{the name of the variable the covariate that will be decomposed}

\item{ant}{the name of the variable that will be used as the ANT variable}

\item{nat}{the name of the variable that will be used as the NAT variable}

\item{time}{the name of the variable that will be used as the time}

\item{hnat}{the name of the variable that indicates wheters the data are from
an historical only nat run.}

\item{knots}{the numbers of knots wanted for the spline in the gam fit}

\item{fixed}{wether the numbers of knots is allowed to vary or not}

\item{correct_ant_bias}{whether the variables that are anthropogenic need to
have their mean shifted to zero in the pre-industrial.}

\item{pre_ind}{period of the pre-industrial if the anthropogenic variables need to be
shifted.}

\item{reuse_ant_bias}{wheter only the shift computed on the first bootstrap
samples is used to reshift the other bootstrap samples. Otherwise each
bootstrap samples is correctied by its own shift.}

\item{all}{the name of the variable that will be used as the ALL variable}

\item{ghg}{the name of the variable that will be used as the GHG variable}

\item{others}{the name of the variable that will be used as the OTHERS variable}
}
\value{
two list :
\itemize{ 
\item bsamples a list of data.frame containing the bootstrap samples with the
decomposition of the covariate x into x_all, x_nat and x_ant
\item osamples a list of data.frame containing the completed/original
datasets for each simulation of the EBM as well as the decomposition of the
covariate x into x_all, x_nat and x_ant
}
}
\description{
\code{dx.gam_allnat} uses the following  gam model x_all = beta_nat * nat + s(ant)

\code{dx.raw} keeps the simulated EBM forcings as it is

\code{dx.lm_gno} uses the follwing linear model : x_all = beta_nat * nat + beta_ghg * ghg + beta_others * others

\code{dx.gam_gno} uses the follwing linear model : gam model x_all = beta_nat * nat + beta_ghg * ghg + s(others)

\code{dx.gam_gno} fits EBM using ordinary least squares for the decomposition
}
