% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebm_sim.r
\name{ebm_bsamples.default}
\alias{ebm_bsamples.default}
\alias{ebm_bsamples.sf_gaussian}
\alias{ebm_bsamples.cm_avg.sf_default}
\title{Generate a dataset of climate response to forcings}
\usage{
ebm_bsamples.default(mdata, model, by = "year", R, gno2aan = TRUE)

ebm_bsamples.sf_gaussian(mdata, model, by = "year", R, gno2aan = TRUE)

ebm_bsamples.cm_avg.sf_default(mdata, model, by = "year", R, gno2aan = TRUE)
}
\arguments{
\item{mdata}{the data to which the EBM simulations will be added}

\item{model, }{the name of the GCM for which the climate responses are needed}

\item{by}{the name of the time variable in mdata. Used to merge mdata with
the EBM response.}

\item{R}{the number of bootstrap simulations}

\item{gno2aan, }{whether to express the climate response into the ALL, ANT and
NAT responses instead of GHG, NAT, and OTHERS(anthropogenic) response}
}
\value{
a list of data.frame where the simulated EBM response are merge to
the dataset mdata
}
\description{
\code{ebm_bsamples.default} generates the climate responses to external
forcings. No random scaling factors are included. The EBM takes the
corresponding GCM parameters if available, otherwise it takes an available
set of paramters chosent at random

\code{ebm_bsamples.sf_gaussian} generates the climate responses to external
forcings with no random scaling factors. it takes the corresponding GCM
parameters if available, otherwise it takes the average of the availables
parameters

\code{ebm_bsamples.default} generates the climate responses to external
forcings. Random gaussian scaling factors on the three categories of forcing
(GHG, NAT, and OTHERS ) are included. The EBM takes the
corresponding GCM parameters if available, otherwise it takes an available
set of paramters chosent at random
}
\examples{
#Simulate 5 EBM realisations and express the results in ALL, ANT and NAT
#responses. In this case, all the EBM simulations are the same since random
#scaling factor are not accounted for and because the EBM paramters for the
#CNRM are presents
data(cnrm)
ebm_bsamples.default(mdata=cnrm, R=5, model="cnrm", by="year", gno2aan=TRUE)
# Simulate 5 EBM realisations with no random scaling factors and express the
# results in GHG, NAT and OTHERS responses
data(cnrm)
ebm_bsamples.sf_gaussian(mdata=cnrm, R=5, model="cnrm", by="year", gno2aan=FALSE)
# Simulate 5 EBM realisations with random scaling factors and express the
# results in GHG, NAT and OTHERS responses
data(cnrm)
ebm_bsamples.sf_gaussian(mdata=cnrm, R=5, model="cnrm", by="year", gno2aan=FALSE)
}
