% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARallnat.r
\docType{package}
\name{FARallnat}
\alias{FARallnat}
\alias{FARallnat-package}
\title{FARallnat  : A Toolbox package for the Analysis of the FAR on the European
Summer temperature.}
\description{
The aims is explain the distribution changes of a variable y with respect to
variations of a variable x
}
\details{
The FAR analysis is decompose in three big steps:
}
\section{Simulations with an EBM of the climate responses}{
 

The simulations are done with respect to three categories of
forcings : anthropogenic(ANT), natural(NAT), and others(OTHERS)
Three variantes are implemented :
\itemize{
\item \code{emb_samples.default}
\item \code{emb_samples.sf_gaussian}
\item \code{emb_samples.cm_avg.sf_default}
}
}

\section{Decomposition of the variable x into ALL,  ANT and NAT}{


Most of the times the EBM simulations help for the decomposition.
several variantes are implemented :
\itemize{
\item \code{dx.gam_allnat}
\item \code{dx.gam_gno}
\item \code{dx.lm_gno}
\item \code{dx.raw}
\item \code{dx.ebm_fit}
}
}

\section{Fit a statistical model to y}{


The model explains variations in the distributions of y with with respect to
the covariate x. From this fit quantities such as the FAR and
the probilities of exceedance can be computed. For the fit, we use three
functions of the FARg package
\itemize{
\item \code{gauss_fit}
\item \code{gev_fit}
\item \code{gpd_fit}
} 
The foo functions ...
}

