% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF.r
\docType{data}
\name{FF}
\alias{FF}
\alias{ebm_params}
\title{Set of parameters and forcings to be used with the EBM model.}
\format{A data frame}
\usage{
FF

ebm_params
}
\description{
Set of parameters in ebm_params and forcings in FF to be used with the EBM model.
}
\keyword{data}
\keyword{datasets}
