% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{format_data}
\alias{format_data}
\title{Function to prepare the tas datasets contained in the package}
\usage{
format_data(tas)
}
\arguments{
\item{tas}{: one of tas dataset included in the package}
}
\description{
\code{format_data} turns tas data into 1961-1991 anomalies
}
