% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pq.r
\name{get_ic_onperiod}
\alias{get_ic_onperiod}
\title{Compute confidence intervals from the bootstrap samples}
\usage{
get_ic_onperiod(b_onperiod, method_name = NULL, ci_p = 0.95, ...)
}
\arguments{
\item{b_onperiod}{array containing the estimates from the bootstrap samples.}

\item{method_name}{adds a columns method to the results with value given by
method_name}

\item{ci_p}{the level of confidence}

\item{...}{additional parameters for the aggregate function use to compute
the bootstrap confidence intervales
array}
}
\value{
returns a data.frame with columns, method, param, IC_inf, Estim (best
estimate), IC_sup .
}
\description{
\code{get_ic_onperiod} Compute the confidences intervals for each parameters
of the bootstrap samples and at each time.
Compute the confidences intervals at confidence level  ci_p for each parameters
of the bootstrap samples and at each time. By default, it computes
90%-confidence intervals. The best estimate corresponds to the median of
bootstrap estimates.
}
\examples{
ans <- compute_far.default(model="cnrm", y="eur_tas", x="gbl_tas", time="year", xp=1.6, stat_model=gauss_fit, ci_p=0.9)
# get bootstrap samples of p_all p_nat and p_ant
bp <- ans$allp
# compute the evolution of p_all relatively to its value in 1850
bp  <- add_param(bp, operation=p_all/p_all[time==1850], name="p_all_rel")
# create the 0.95 confidence intervals data.frame
ic  <- get_ic_onperiod(bp, ci=0.95)
# To be done
}
