% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_fun.r
\name{gg_color_hue}
\alias{gg_color_hue}
\title{generate a vactor of n equally spaced (in terms of hue)  colors.}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{the number of color wanted}
}
\value{
a vector of colors
}
\description{
\code{gg_color_hue} generate a vector of n equally spaced (in terms of hue)
colorsi for ggplot2 plots.
}
\examples{
gg_color_hue(10)
}

