% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_dual_axis.r
\name{ggplot_dual_axis}
\alias{ggplot_dual_axis}
\title{function named ggplot_dual_axis()
Takes 2 ggplot plots and makes a dual y-axis plot
function takes 2 compulsory arguments and 1 optional argument
arg lhs is the ggplot whose y-axis is to be displayed on the left
arg rhs is the ggplot whose y-axis is to be displayed on the right
arg 'axis.title.y.rhs' takes value "rotate" to rotate right y-axis label
The function does as little as possible, namely:
 # display the lhs plot without minor grid lines and with a
 transparent background to allow grid lines to show
 # display the rhs plot without minor grid lines and with a
 secondary y axis, a rotated axis label, without minor grid lines
 # justify the y-axis label by setting 'hjust = 0' in 'axis.text.y'
 # rotate the right plot 'axis.title.y' by 270 degrees, for symmetry
 # rotation can be turned off with 'axis.title.y.rhs' option}
\usage{
ggplot_dual_axis(lhs, rhs, axis.title.y.rhs = "rotate")
}
\description{
function named ggplot_dual_axis()
Takes 2 ggplot plots and makes a dual y-axis plot
function takes 2 compulsory arguments and 1 optional argument
arg lhs is the ggplot whose y-axis is to be displayed on the left
arg rhs is the ggplot whose y-axis is to be displayed on the right
arg 'axis.title.y.rhs' takes value "rotate" to rotate right y-axis label
The function does as little as possible, namely:
 # display the lhs plot without minor grid lines and with a
 transparent background to allow grid lines to show
 # display the rhs plot without minor grid lines and with a
 secondary y axis, a rotated axis label, without minor grid lines
 # justify the y-axis label by setting 'hjust = 0' in 'axis.text.y'
 # rotate the right plot 'axis.title.y' by 270 degrees, for symmetry
 # rotation can be turned off with 'axis.title.y.rhs' option
}

