% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pq.r
\name{imput_aurel}
\alias{imput_aurel}
\title{Imputs value in the case of Relative Risk(RR) being equal to NA}
\usage{
imput_aurel(boot_res_RR)
}
\arguments{
\item{boot_res_RR}{a matrix of the bootstrap estimate of RR for each year and
each bootstrap sample}
}
\value{
returns a matrix the NA value replaced half of the time by +infinity
and half by the time by zero
}
\description{
\code{imput_aure} Imputs value in the case of Relative Risk(RR) being equal
to NA for each bootstrap sample and for each year.
}
\examples{
# Needs to be done !
}
