% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fun.r
\name{plot_boot_time}
\alias{plot_boot_time}
\title{\code{plot_far} Plot the times series of one model parameters and its
confidence intervales}
\usage{
plot_boot_time(ic_df, param = "FAR", main = "", alpha = 0.2)
}
\arguments{
\item{ic_df}{a data_frame contaning the parameter estimates and their confidence
intervals}

\item{param}{the name of the parameter to be plotted}

\item{main}{the title of the plot}

\item{alpha}{the transparency of the shading for the confidence intervals}
}
\description{
\code{plot_far} Plot the times series of one model parameters and its
confidence intervales
}
\examples{
library(FARg)
model <- "cnrm"
#load data from the package
data(list=model)
# formating data, e.g passing from temperature to anomalie, keep only hist
# and rcp runs
mdata <- format_data(get(model))
if(model != "obs") mdata <- select_continuous_run(mdata)
ans <- compute_far_simple.default(mdata,
 y="eur_tas", x="gbl_tas", time="year",
  xp=1.6, stat_model=gauss_fit, ci_p=0.9)
plot_boot_time(ans$ic_far, param="p_all", main="p_all") 
}
