% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fun.r
\name{plot_far}
\alias{plot_far}
\title{\code{plot_far} create a symetric double y-axis plot for the evolution FAR and the RR}
\usage{
plot_far(ic_mat, axis_trans, main = "", xlim, ylim, col = NULL)
}
\arguments{
\item{ic_mat}{a data_frame contaning the RR estimates and their confidence
intervales and the rescaled value of RR (either RRal or RRel, see axis_trans)}

\item{axis_trans}{the choice of doublei : axis = "al" for a symetric
atan(log(y)) axis ; axis_trans = "el" for a symetrized FAR axis}

\item{main}{the title of the plot}

\item{xlim}{a vector of two values for the limits of the x-axis}

\item{ylim}{a vector of two values for the limits of the y-axis}

\item{col}{a vector of colors for custom colors, should be the same length
as the number of different values of the ic_mat method column.}
}
\description{
\code{plot_far} create a symetric double y-axis plot for the evolution FAR and the RR
}
\examples{
library(FARg)
#load data from the package
# formating data, e.g passing from temperature to anomalie, keep only hist
# and rcp runs
mdata <- format_data(cnrm)
# keep only run that posses rcp simulations
mdata <- select_continuous_run(mdata)
ans <- compute_far_simple(mdata,
 y="eur_tas", x="gbl_tas", time="year",
  xp=1.6, stat_model=gauss_fit, ci_p=0.9)
plot_far(ans$ic_far, axis_trans="al", col=gg_color_hue(2))
}
