% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fun.r
\name{reorder_layer}
\alias{reorder_layer}
\title{\code{reorder_layer} Put the bottom layer of a ggplot on top}
\usage{
reorder_layer(gp)
}
\arguments{
\item{gp}{a ggplot2 plot}
}
\description{
\code{reorder_layer} Put the bottom layer of a ggplot on top
}
\examples{
library(FARg)
#load data from the package
# formating data, e.g passing from temperature to anomalie, keep only hist
# and rcp runs
mdata <- format_data(cnrm)
# keep only run that posses rcp simulations
mdata <- select_continuous_run(mdata)
ans <- compute_far_simple(mdata,
 y="eur_tas", x="gbl_tas", time="year",
  xp=1.6, stat_model=gauss_fit, ci_p=0.9)
plot_far(ans$ic_far, axis_trans="al", col=gg_color_hue(2))
}
