% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{select_continuous_run}
\alias{select_continuous_run}
\title{Function to prepare the tas datasets contained in the package}
\usage{
select_continuous_run(tas)
}
\arguments{
\item{tas}{: one of tas dataset included in the package}
}
\description{
\code{select_continuous_run} keeps only climate runs with historical and rcp
}
