% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.r
\name{summary_plot}
\alias{summary_plot}
\title{Summary plots of the estimated FAR}
\usage{
summary_plot(merged_res, pdf_name = "summary.pdf")
}
\arguments{
\item{merged_res}{results of the compute_far_fonction}
}
\value{
saves a pdf files called "summary.pdf" with the plots in same folder
as the res_forlder of the compute_forall function
}
\description{
\code{summary_plot} computes a fex plots  to summarize the results of
compute_forall
}
\examples{
# compute the FAR for the CNRM and the IPSL GCMs using a  gam decomposition
# and a gaussian fit with only three bootstrap samples
ans <- compute_forall(compute_far.default, stat_model=gauss_fit, p=0.01, R=3)
# summary plot
summary_plot(ans)
}

