% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_fun.r
\name{plot_boot_time}
\alias{plot_boot_time}
\title{\code{plot_far} Plot the times series of one model parameters and its
confidence intervales}
\usage{
plot_boot_time(ic_df, param = "FAR", main = "", alpha = 0.2)
}
\arguments{
\item{ic_df}{a data_frame contaning the parameter estimates and their confidence
intervals}

\item{param}{the name of the parameter to be plotted}

\item{main}{the title of the plot}

\item{alpha}{the transparency of the shading for the confidence intervals}
}
\description{
\code{plot_far} Plot the times series of one model parameters and its
confidence intervales
}
\examples{
ans <- compute_forall(compute_far.dx_ebm_fit, l_models=c("cnrm", "ipsl"), stat_model=gpd_fit, qthreshold =0.90, p=0.01, R=3)
plot_boot_time(ans, param="p_all", main="p_all")
}

